✨ feat: QueryService, CommandService 분리
