+++ src/main/java/com/chatsul/service/LostItemCommandService.java
@@ -1,0 +1,11 @@
+package com.chatsul.service;
+
+import com.chatsul.web.dto.LostItemRequestDTO;
+
+public interface LostItemCommandService {
+    String saveLostItem(LostItemRequestDTO lostItemRequestDTO);
+
+    void deleteLostItem(LostItemRequestDTO lostItemRequestDTO);
+
+    String updateLostItemStatus(Long lostItemId);
+}
